feat(backend): :sparkles: create get_my_teaching api
I refactor router and then Implemnt get_my_teaching
