+++ backend/backend/main.py
@@ -6,31 +6,23 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from backend.config import APP_NAME, API_HOST, API_PORT, LOG_LEVEL
 from backend.routers import (
-    course_create,
-    example,
-    info,
     authentication,
     review,
     buy_course,
     course,
     category,
-    user,
     search,
-    add_course_to_cart,
     list_course_on_home_page,
     list_everything,
+    user,
 )
 
 app = FastAPI()
 
-app.include_router(example.router)
-app.include_router(info.router)
 app.include_router(authentication.router)
 app.include_router(search.router)
-app.include_router(add_course_to_cart.router)
 app.include_router(list_course_on_home_page.router)
 app.include_router(list_everything.router)
-app.include_router(course_create.router)
 # app.include_router(view_my_learning.router)
 app.include_router(review.router)
 app.include_router(course.router)
