+++ backend/backend/routers/cart.py
@@ -11,13 +11,6 @@ from backend.lib.authentication import get_current_user
 
 router = APIRouter()
 
-# @router.get("/user/{user_name}")
-# def get_user(user_name: str):
-#     return controller.get_users_by_name(user_name)
-
-# @router.get("/course/{course_name}")
-# def get_course(course_name: str):
-#     return controller.search_course_by_name(course_name)
 route_tags: List[str | Enum] = ["cart"]
     
 @router.post('/user/add_course_to_cart')
@@ -41,6 +34,4 @@ def remove_course_to_cart(current_user: Annotated[User, Depends(get_current_user
 
 @router.get('/user/{user_id}/cart/')
 def get_course_in_cart(current_user: Annotated[User, Depends(get_current_user)], course_id_query: UUID):
-
-
     return current_user.get_cart()
