+++ backend/backend/routers/course.py
@@ -1,21 +1,30 @@
 from enum import Enum
 from typing import Annotated, List
 import uuid
-from fastapi import APIRouter, Depends, HTTPException, status
-
+from fastapi import APIRouter, Depends, HTTPException, status, Body, Response
 from backend.controller_instance import controller
-from backend.definitions.user import User
+from backend.definitions.user import User,Teacher
 from backend.definitions.api_data_model import CourseInfo, CourseLearn, CourseLearnMaterialImage, CourseLearnMaterialQuiz, CourseLearnMaterialQuizQuestions, CourseLearnMaterialVideo
 from backend.lib.authentication import get_current_user
+from pydantic import BaseModel
+
+from backend.definitions.course import (
+    Course,
+    CourseMaterialImage,
+    CourseMaterialQuiz,
+    QuizQuestion,
+)
 
 
 router = APIRouter()
-route_tags: List[str | Enum] = ["course"]
+route_tags: List[str | Enum] = ["Course"]
 
 
 @router.get("/course", tags=route_tags)
 def get_all_course():
     return controller.get_all_courses()
+
+
 
 
 @router.get("/course/{course_id}", tags=route_tags)
@@ -132,3 +141,149 @@ def get_learn_course_materials(
     )
 
     return course_learn_data
+
+
+class PostCourseData(BaseModel):
+    name: str
+    description: str
+    price: int
+    category_id: str
+
+
+@router.post("/course", tags=route_tags)
+def new_course(
+    post_course_data: Annotated[
+        PostCourseData,
+        Body(
+            examples=[
+                {
+                    "name": "Example",
+                    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+                    "price": 100,
+                    "category_id": "afeabc96-80f6-4508-82e4-e8a429e86547",
+                }
+            ],
+        ),
+    ],
+    response: Response,
+    current_user: Annotated[User, Depends(get_current_user)],
+):
+    teacher = current_user
+
+    if not isinstance(teacher, Teacher):
+        response.status_code = status.HTTP_400_BAD_REQUEST
+        return "Teacher not found"
+
+    category = controller.search_category_by_id(uuid.UUID(post_course_data.category_id))
+    if not category:
+        response.status_code = status.HTTP_400_BAD_REQUEST
+        return "Category not found"
+
+    course = Course(
+        post_course_data.name,
+        post_course_data.description,
+        post_course_data.price,
+        current_user,
+    )
+
+    teacher.add_my_teaching(course)
+    category.add_course(course)
+
+    return course
+
+
+class CourseMaterialPostData(BaseModel):
+    name: str
+    description: str
+
+
+class AddImageToCoursePostData(CourseMaterialPostData):
+    url: str
+
+
+
+@router.post("/course/{course_id}/image", tags=route_tags)
+def add_image_to_course(
+    course_id: str,
+    add_image_to_course_data: Annotated[
+        AddImageToCoursePostData,
+        Body(
+            examples=[
+                {
+                    "url": "https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png",
+                    "name": "FastAPI Logo",
+                    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+                }
+            ],
+        ),
+    ],
+    response: Response,
+    current_user: Annotated[User, Depends(get_current_user)],
+):
+    course = controller.search_course_by_id(uuid.UUID(course_id))
+    if not isinstance(course, Course):
+        response.status_code = status.HTTP_400_BAD_REQUEST
+        return "Course not found"
+
+    if not (current_user == course.get_teacher()):
+        response.status_code = status.HTTP_401_UNAUTHORIZED
+        return "Unauthorized"
+
+    image = CourseMaterialImage(
+        add_image_to_course_data.url,
+        add_image_to_course_data.name,
+        add_image_to_course_data.description,
+    )
+    course.add_image(image)
+
+    return course
+
+
+class QuizQuestionPostData(BaseModel):
+    question: str
+    correct: bool
+
+
+class AddQuizToCoursePostData(CourseMaterialPostData):
+    questions: List[QuizQuestionPostData]
+
+
+@router.post("/course/{course_id}/quiz", tags=route_tags)
+def add_quiz_to_course(
+    course_id: str,
+    add_quiz_to_course_data: Annotated[
+        AddQuizToCoursePostData,
+        Body(
+            examples=[
+                {
+                    "name": "Which language is FastAPI built with",
+                    "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+                    "questions": [
+                        {"question": "Python", "correct": True},
+                        {"question": "Rust", "correct": False},
+                    ],
+                }
+            ],
+        ),
+    ],
+    response: Response,
+    current_user: Annotated[User, Depends(get_current_user)],
+):
+    course = controller.search_course_by_id(uuid.UUID(course_id))
+    if not isinstance(course, Course):
+        response.status_code = status.HTTP_400_BAD_REQUEST
+        return "Teacher not found"
+
+    if not (current_user == course.get_teacher()):
+        response.status_code = status.HTTP_401_UNAUTHORIZED
+        return "Unauthorized"
+
+    quiz = CourseMaterialQuiz(
+        add_quiz_to_course_data.name, add_quiz_to_course_data.description
+    )
+    for question in add_quiz_to_course_data.questions:
+        quiz.add_question(QuizQuestion(question.question, question.correct))
+    course.add_quiz(quiz)
+
+    return course
+
