--- backend/backend/routers/example.py
@@ -1,8 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter()
-
-
-@router.get("/example/")
-async def get_example():
-    return "hi"
