+++ backend/backend/routers/list_everything.py
@@ -1,7 +1,7 @@
 from typing import List, Union
 from fastapi import APIRouter, Body
 from pydantic import UUID4
-
+from enum import Enum
 from backend.definitions.course import Course
 from backend.definitions.user import User, Teacher
 from backend.definitions.controller import Controller
@@ -10,17 +10,22 @@ import random
 
 router = APIRouter()
 
-@router.get("/course/all_course/")
+route_tags: List[str | Enum] = ["Debug Purposes"]
+
+@router.get("/course/all_course/", tags = route_tags)
 def get_all_course_list():
     all_courses = controller.get_all_courses()
     return all_courses
 
-@router.get("/teacher/all_teacher")
+@router.get("/teacher/all_teacher", tags = route_tags)
 def get_all_teacher_list():
     return controller.get_all_teacher()
 
-@router.get("/user/all_user")
+
+#for debug only
+@router.get("/user", tags=route_tags)
 def get_all_user():
     return controller.get_all_user()
+
