--- backend/backend/routers/order.py
@@ -1,9 +1,0 @@
-from fastapi import APIRouter, Body
-from backend.definitions.order import Order,Payment
-
-router = APIRouter()
-
-@router.get("/examplerrrr/{status}{amount}{country}{user}{payment_method}")
-async def createpayment(status):
-    return "hi"
-
